Changes made at meeting 12/5
